refactor: fixed the imports in testing files to match new file structure
