+++ app/backend/testing/integration/test_guest_live_map_integration.py
@@ -1,8 +1,8 @@
 import pytest
 from geoalchemy2.elements import WKTElement
 
-from enums.report_status import ReportStatus
-from models.reported_fires import FireReports
+from src.enums.report_status import ReportStatus
+from src.models.reported_fires import FireReports
 
 
 @pytest.mark.skip(reason="PostGIS not configured for testing yet.")
