+++ app/backend/testing/integration/test_xgboost_model_integration.py
@@ -8,9 +8,9 @@ from ml.training.synthetic_data import (
 )
 from ml.training.train_ignition import group_split, train
 
-from ai import artifact_store
-from ai.ignition import IgnitionScorer
-from ai.schema import BURNED, BURNING, FEATURES, SCHEMA_VERSION, UNBURNED
+from src.ai import artifact_store
+from src.ai.ignition import IgnitionScorer
+from src.ai.schema import BURNED, BURNING, FEATURES, SCHEMA_VERSION, UNBURNED
 
 
 # Shared fixture. (scope="module" -> Trains once per pytest run)
