+++ app/backend/testing/unit/test_xgboost_model_unit.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pytest
 
-from ai.features import grid_to_fmatrix, neighbour_features, shift
-from ai.schema import (
+from src.ai.features import grid_to_fmatrix, neighbour_features, shift
+from src.ai.schema import (
     BURNED,
     BURNING,
     DIST_CAP,
